+++ execution_engine/src/core/runtime_context/mod.rs
@@ -329,6 +329,14 @@ where
                 error!("should not remove the chainspec registry key");
                 Err(Error::RemoveKeyFailure(RemoveKeyFailure::PermissionDenied))
             }
+            Key::BlockEffectsRootHash { .. } => {
+                error!("should not remove the execution results root hash key");
+                Err(Error::RemoveKeyFailure(RemoveKeyFailure::PermissionDenied))
+            }
+            Key::DeployApprovalsRootHash { .. } => {
+                error!("should not remove the deploy approvals root hash key");
+                Err(Error::RemoveKeyFailure(RemoveKeyFailure::PermissionDenied))
+            }
         }
     }
 
@@ -785,6 +793,8 @@ where
             Key::Dictionary(_) => true,
             Key::SystemContractRegistry => true,
             Key::ChainspecRegistry => true,
+            Key::BlockEffectsRootHash { .. } => true,
+            Key::DeployApprovalsRootHash { .. } => true,
         }
     }
 
@@ -807,6 +817,8 @@ where
             }
             Key::SystemContractRegistry => false,
             Key::ChainspecRegistry => false,
+            Key::BlockEffectsRootHash { .. } => false,
+            Key::DeployApprovalsRootHash { .. } => false,
         }
     }
 
@@ -829,6 +841,8 @@ where
             }
             Key::SystemContractRegistry => false,
             Key::ChainspecRegistry => false,
+            Key::BlockEffectsRootHash { .. } => false,
+            Key::DeployApprovalsRootHash { .. } => false,
         }
     }