+++ node/src/components/contract_runtime.rs
@@ -41,7 +41,7 @@ use casper_global_state::{
     },
 };
 use casper_hashing::Digest;
-use casper_types::{bytesrepr::Bytes, EraId, ProtocolVersion, Timestamp};
+use casper_types::{bytesrepr::Bytes, ProtocolVersion, Timestamp};
 
 use crate::{
     components::{contract_runtime::types::StepEffectAndUpcomingEraValidators, Component},
@@ -147,14 +147,11 @@ impl ExecutionPreState {
 
     /// Creates instance of `ExecutionPreState` from given block header nad merkle tree hash
     /// activation point.
-    pub fn from_block_header(
-        block_header: &BlockHeader,
-        verifiable_chunked_hash_activation: EraId,
-    ) -> Self {
+    pub fn from_block_header(block_header: &BlockHeader) -> Self {
         ExecutionPreState {
             pre_state_root_hash: *block_header.state_root_hash(),
             next_block_height: block_header.height() + 1,
-            parent_hash: block_header.hash(verifiable_chunked_hash_activation),
+            parent_hash: block_header.hash(),
             parent_seed: block_header.accumulated_seed(),
         }
     }
@@ -199,7 +196,6 @@ pub(crate) struct ContractRuntime {
     engine_state: Arc<EngineState<LmdbGlobalState>>,
     metrics: Arc<Metrics>,
     protocol_version: ProtocolVersion,
-    verifiable_chunked_hash_activation: EraId,
 
     /// Finalized blocks waiting for their pre-state hash to start executing.
     exec_queue: ExecQueue,
@@ -499,7 +495,6 @@ impl ContractRuntime {
                 );
                 let engine_state = Arc::clone(&self.engine_state);
                 let metrics = Arc::clone(&self.metrics);
-                let verifiable_chunked_hash_activation = self.verifiable_chunked_hash_activation();
                 async move {
                     let result = run_intensive_task(move || {
                         execute_finalized_block(
@@ -510,7 +505,6 @@ impl ContractRuntime {
                             finalized_block,
                             deploys,
                             transfers,
-                            verifiable_chunked_hash_activation,
                         )
                     })
                     .await;
@@ -545,7 +539,6 @@ impl ContractRuntime {
                             finalized_block,
                             deploys,
                             transfers,
-                            self.verifiable_chunked_hash_activation(),
                         )
                         .ignore(),
                     )
@@ -626,7 +619,6 @@ impl ContractRuntime {
         strict_argument_checking: bool,
         vesting_schedule_period_millis: u64,
         registry: &Registry,
-        verifiable_chunked_hash_activation: EraId,
     ) -> Result<Self, ConfigError> {
         // TODO: This is bogus, get rid of this
         let execution_pre_state = Arc::new(Mutex::new(ExecutionPreState {
@@ -670,14 +662,9 @@ impl ContractRuntime {
             engine_state,
             metrics,
             protocol_version,
-            verifiable_chunked_hash_activation,
             exec_queue: Arc::new(Mutex::new(BTreeMap::new())),
             system_contract_registry: None,
         })
-    }
-
-    fn verifiable_chunked_hash_activation(&self) -> EraId {
-        self.verifiable_chunked_hash_activation
     }
 
     /// Commits a genesis request.
@@ -790,7 +777,6 @@ impl ContractRuntime {
         finalized_block: FinalizedBlock,
         deploys: Vec<Deploy>,
         transfers: Vec<Deploy>,
-        verifiable_chunked_hash_activation: EraId,
     ) where
         REv: From<ContractRuntimeRequest>
             + From<ContractRuntimeAnnouncement>
@@ -812,7 +798,6 @@ impl ContractRuntime {
                 finalized_block,
                 deploys,
                 transfers,
-                verifiable_chunked_hash_activation,
             )
         })
         .await
@@ -821,10 +806,7 @@ impl ContractRuntime {
             Err(error) => return fatal!(effect_builder, "{}", error).await,
         };
 
-        let new_execution_pre_state = ExecutionPreState::from_block_header(
-            block.header(),
-            verifiable_chunked_hash_activation,
-        );
+        let new_execution_pre_state = ExecutionPreState::from_block_header(block.header());
         *execution_pre_state.lock().unwrap() = new_execution_pre_state.clone();
 
         let current_era_id = block.header().era_id();