+++ node/src/reactor/joiner.rs
@@ -565,10 +565,6 @@ impl reactor::Reactor for Reactor {
             sync_effects,
         ));
 
-        let verifiable_chunked_hash_activation = chainspec_loader
-            .chainspec()
-            .protocol_config
-            .verifiable_chunked_hash_activation;
         let protocol_version = &chainspec_loader.chainspec().protocol_config.version;
         let rest_server = RestServer::new(
             config.rest_server.clone(),
@@ -583,8 +579,7 @@ impl reactor::Reactor for Reactor {
             *protocol_version,
         )?;
 
-        let fetcher_builder =
-            FetcherBuilder::new(config.fetcher, registry, verifiable_chunked_hash_activation);
+        let fetcher_builder = FetcherBuilder::new(config.fetcher, registry);
 
         let deploy_fetcher = fetcher_builder.build("deploy")?;
         let finalized_approvals_fetcher = fetcher_builder.build("finalized_approvals")?;
@@ -599,7 +594,7 @@ impl reactor::Reactor for Reactor {
 
         let trie_or_chunk_fetcher = fetcher_builder.build("trie_or_chunk")?;
 
-        let deploy_acceptor = DeployAcceptor::new(&*chainspec_loader.chainspec(), registry)?;
+        let deploy_acceptor = DeployAcceptor::new(chainspec_loader.chainspec(), registry)?;
 
         let deploy_gossiper = Gossiper::new_for_partial_items(
             "deploy_gossiper",
@@ -693,7 +688,6 @@ impl reactor::Reactor for Reactor {
                 ));
 
                 let event = fetcher::Event::GotRemotely {
-                    verifiable_chunked_hash_activation: None,
                     item: deploy,
                     source,
                 };
@@ -954,19 +948,7 @@ impl reactor::Reactor for Reactor {
                     .handle_event(effect_builder, rng, incoming.into()),
             ),
             JoinerEvent::NetResponseIncoming(NetResponseIncoming { sender, message }) => {
-                let verifiable_chunked_hash_activation = self
-                    .chainspec_loader
-                    .chainspec()
-                    .protocol_config
-                    .verifiable_chunked_hash_activation;
-                reactor::handle_get_response(
-                    self,
-                    effect_builder,
-                    rng,
-                    sender,
-                    message,
-                    verifiable_chunked_hash_activation,
-                )
+                reactor::handle_get_response(self, effect_builder, rng, sender, message)
             }
             JoinerEvent::TrieRequestIncoming(incoming) => reactor::wrap_effects(
                 JoinerEvent::ContractRuntime,
@@ -985,10 +967,6 @@ impl reactor::Reactor for Reactor {
                     rng,
                     sender,
                     &message.0,
-                    self.chainspec_loader
-                        .chainspec()
-                        .protocol_config
-                        .verifiable_chunked_hash_activation,
                 )
             }