+++ node/src/types/chainspec/protocol_config.rs
@@ -7,13 +7,12 @@ use datasize::DataSize;
 #[cfg(test)]
 use rand::Rng;
 use serde::{Deserialize, Serialize};
-use tracing::error;
 
 #[cfg(test)]
 use casper_types::testing::TestRng;
 use casper_types::{
     bytesrepr::{self, FromBytes, ToBytes},
-    EraId, Key, ProtocolVersion, StoredValue,
+    Key, ProtocolVersion, StoredValue,
 };
 
 use super::{ActivationPoint, GlobalStateUpdate};
@@ -31,10 +30,6 @@ pub struct ProtocolConfig {
     /// Any arbitrary updates we might want to make to the global state at the start of the era
     /// specified in the activation point.
     pub(crate) global_state_update: Option<GlobalStateUpdate>,
-    /// The era ID in which the last emergency restart happened.
-    pub(crate) last_emergency_restart: Option<EraId>,
-    /// The era ID starting at which the new Merkle tree-based hashing scheme is applied.
-    pub(crate) verifiable_chunked_hash_activation: EraId,
 }
 
 impl ProtocolConfig {
@@ -65,20 +60,6 @@ impl ProtocolConfig {
 
     /// Checks whether the values set in the config make sense and returns `false` if they don't.
     pub(super) fn is_valid(&self) -> bool {
-        // Assert the `last_emergency_restart` is `None` or no more than `activation_point`.
-        if let Some(last_emergency_restart) = self.last_emergency_restart {
-            let activation_point = self.activation_point.era_id();
-            if last_emergency_restart > activation_point {
-                error!(
-                    %activation_point,
-                    %last_emergency_restart,
-                    "[protocol.last_emergency_restart] cannot be greater than \
-                    [protocol.activation_point] in the chainspec."
-                );
-                return false;
-            };
-        }
-
         true
     }
 
@@ -91,16 +72,12 @@ impl ProtocolConfig {
             rng.gen::<u8>() as u32,
         );
         let activation_point = ActivationPoint::random(rng);
-        let last_emergency_restart = rng.gen::<bool>().then(|| rng.gen());
-        let verifiable_chunked_hash_activation = EraId::from(rng.gen_range(0..5));
 
         ProtocolConfig {
             version: protocol_version,
             hard_reset: rng.gen(),
             activation_point,
             global_state_update: None,
-            last_emergency_restart,
-            verifiable_chunked_hash_activation,
         }
     }
 }
@@ -112,8 +89,6 @@ impl ToBytes for ProtocolConfig {
         buffer.extend(self.hard_reset.to_bytes()?);
         buffer.extend(self.activation_point.to_bytes()?);
         buffer.extend(self.global_state_update.to_bytes()?);
-        buffer.extend(self.last_emergency_restart.to_bytes()?);
-        buffer.extend(self.verifiable_chunked_hash_activation.to_bytes()?);
         Ok(buffer)
     }
 
@@ -122,8 +97,6 @@ impl ToBytes for ProtocolConfig {
             + self.hard_reset.serialized_length()
             + self.activation_point.serialized_length()
             + self.global_state_update.serialized_length()
-            + self.last_emergency_restart.serialized_length()
-            + self.verifiable_chunked_hash_activation.serialized_length()
     }
 }
 
@@ -135,15 +108,11 @@ impl FromBytes for ProtocolConfig {
         let (hard_reset, remainder) = bool::from_bytes(remainder)?;
         let (activation_point, remainder) = ActivationPoint::from_bytes(remainder)?;
         let (global_state_update, remainder) = Option::<GlobalStateUpdate>::from_bytes(remainder)?;
-        let (last_emergency_restart, remainder) = Option::<EraId>::from_bytes(remainder)?;
-        let (verifiable_chunked_hash_activation, remainder) = EraId::from_bytes(remainder)?;
         let protocol_config = ProtocolConfig {
             version,
             hard_reset,
             activation_point,
             global_state_update,
-            last_emergency_restart,
-            verifiable_chunked_hash_activation,
         };
         Ok((protocol_config, remainder))
     }
@@ -152,6 +121,8 @@ impl FromBytes for ProtocolConfig {
 #[cfg(test)]
 mod tests {
     use crate::types::Block;
+
+    use casper_types::EraId;
 
     use super::*;
 
@@ -183,18 +154,10 @@ mod tests {
         let mut rng = crate::new_rng();
         let mut protocol_config = ProtocolConfig::random(&mut rng);
 
-        let activation_point = EraId::new(rng.gen_range(2..u64::MAX));
-
         // We force `global_state_update` to be `None`.
         protocol_config.global_state_update = None;
 
-        // The config is valid only if `last_emergency_restart` is no more than `activation_point`.
-        protocol_config.activation_point = ActivationPoint::EraId(activation_point);
-        protocol_config.last_emergency_restart = Some(activation_point);
         assert!(protocol_config.is_valid());
-
-        protocol_config.last_emergency_restart = Some(activation_point + 1);
-        assert!(!protocol_config.is_valid());
     }
 
     #[test]
@@ -202,18 +165,10 @@ mod tests {
         let mut rng = crate::new_rng();
         let mut protocol_config = ProtocolConfig::random(&mut rng);
 
-        let activation_point = EraId::new(rng.gen_range(2..u64::MAX));
-
         // We force `global_state_update` to be `Some`.
         protocol_config.global_state_update = Some(GlobalStateUpdate::random(&mut rng));
 
-        // The config is valid only if `last_emergency_restart` is no more than `activation_point`.
-        protocol_config.activation_point = ActivationPoint::EraId(activation_point);
-        protocol_config.last_emergency_restart = Some(activation_point);
         assert!(protocol_config.is_valid());
-
-        protocol_config.last_emergency_restart = Some(activation_point + 1);
-        assert!(!protocol_config.is_valid());
     }
 
     #[test]
@@ -225,74 +180,35 @@ mod tests {
         let upgrade_era = EraId::from(5);
         let previous_era = upgrade_era.saturating_sub(1);
 
-        let verifiable_chunked_hash_activation = EraId::from(0);
-
         let mut rng = crate::new_rng();
         let protocol_config = ProtocolConfig {
             version: current_version,
             hard_reset: false,
             activation_point: ActivationPoint::EraId(upgrade_era),
             global_state_update: None,
-            last_emergency_restart: None,
-            verifiable_chunked_hash_activation,
         };
 
         // The block before this protocol version: a switch block with previous era and version.
-        let block = Block::random_with_specifics(
-            &mut rng,
-            previous_era,
-            100,
-            past_version,
-            true,
-            verifiable_chunked_hash_activation,
-            None,
-        );
+        let block =
+            Block::random_with_specifics(&mut rng, previous_era, 100, past_version, true, None);
         assert!(protocol_config.is_last_block_before_activation(block.header()));
 
         // Not the activation point: wrong era.
-        let block = Block::random_with_specifics(
-            &mut rng,
-            upgrade_era,
-            100,
-            past_version,
-            true,
-            verifiable_chunked_hash_activation,
-            None,
-        );
+        let block =
+            Block::random_with_specifics(&mut rng, upgrade_era, 100, past_version, true, None);
         assert!(!protocol_config.is_last_block_before_activation(block.header()));
 
         // Not the activation point: wrong version.
-        let block = Block::random_with_specifics(
-            &mut rng,
-            previous_era,
-            100,
-            current_version,
-            true,
-            verifiable_chunked_hash_activation,
-            None,
-        );
+        let block =
+            Block::random_with_specifics(&mut rng, previous_era, 100, current_version, true, None);
         assert!(!protocol_config.is_last_block_before_activation(block.header()));
-        let block = Block::random_with_specifics(
-            &mut rng,
-            previous_era,
-            100,
-            future_version,
-            true,
-            verifiable_chunked_hash_activation,
-            None,
-        );
+        let block =
+            Block::random_with_specifics(&mut rng, previous_era, 100, future_version, true, None);
         assert!(!protocol_config.is_last_block_before_activation(block.header()));
 
         // Not the activation point: not a switch block.
-        let block = Block::random_with_specifics(
-            &mut rng,
-            previous_era,
-            100,
-            past_version,
-            false,
-            verifiable_chunked_hash_activation,
-            None,
-        );
+        let block =
+            Block::random_with_specifics(&mut rng, previous_era, 100, past_version, false, None);
         assert!(!protocol_config.is_last_block_before_activation(block.header()));
     }
 }