+++ utils/nctl/sh/scenarios-upgrades/upgrade_scenario_05.sh
@@ -1,4 +1,4 @@
-#!/usr/bin/env bash
+#!/Usr/bin/env bash
 # ----------------------------------------------------------------
 # Synopsis.
 # ----------------------------------------------------------------
@@ -91,9 +91,7 @@ function _step_01()
     log_step_upgrades 0 "Begin upgrade_scenario_05"
     log_step_upgrades 1 "starting network from stage ($STAGE_ID)"
 
-    source "$NCTL/sh/assets/setup_from_stage.sh" \
-            stage="$STAGE_ID" \
-            chainspec_path="$NCTL/overrides/upgrade_scenario_5.pre.chainspec.toml.in"
+    source "$NCTL/sh/assets/setup_from_stage.sh" stage="$STAGE_ID"
     log "... Starting 5 validators"
     source "$NCTL/sh/node/start.sh" node=all
     log "... Starting 5 non-validators"
@@ -130,8 +128,7 @@ function _step_03()
             stage="$STAGE_ID" \
             verbose=false \
             node="$i" \
-            era="$ACTIVATION_POINT" \
-            chainspec_path="$NCTL/overrides/upgrade_scenario_5.post.chainspec.toml.in"
+            era="$ACTIVATION_POINT"
         echo ""
     done
 
@@ -238,7 +235,7 @@ function _step_07()
 function _step_08()
 {
     log_step_upgrades 8 "awaiting 10 eras"
-    await_n_eras '10' 'true' '5.0' '2'
+    nctl-await-n-eras offset='10' sleep_interval='5.0' timeout='600' node_id='2'
 }
 
 # Step 09: Stage nodes 1&10 and restart.
@@ -268,8 +265,7 @@ function _step_09()
             stage="$STAGE_ID" \
             verbose=false \
             node="$i" \
-            era="$ACTIVATION_POINT" \
-            chainspec_path="$NCTL/overrides/upgrade_scenario_5.post.chainspec.toml.in"
+            era="$ACTIVATION_POINT"
         echo ""
         # add hash to upgrades config
         PATH_TO_NODE_CONFIG_UPGRADE="$(get_path_to_node_config $i)/$N2_PROTO_VERSION/config.toml"